Review comment overlay changes
- move the central init method to a separate file in core
- add None return type to a helper protected method
- fix imports
